requirements.txt
langgraph>=0.0.20
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-google-genai>=0.0.3
sqlalchemy>=2.0.23
pandas>=2.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0
redis>=5.0.0
openai>=1.3.7
pydantic>=2.5.0
requests>=2.31.0
geopy>=2.4.0
pytrends>=4.9.2
colorama>=0.4.6
aiohttp>=3.9.0
pyyaml>=6.0
# Email dependencies
aiosmtplib>=2.0.2
jinja2>=3.1.0
langchain-anthropic>=0.1.8

# Added for What-If scenario agent
langchain-community>=0.0.20
duckduckgo-search
pyowm